use simplecov and codeclimate for unit test coverage reports
